Fix PHP 8.1+ float to int conversion deprecation warning
- Cast floor() result to int in format_bytes() function
- Prevents 'Implicit conversion from float to int loses precision' warnings
- Makes code compatible with PHP 8.1+ strict type checking
- Fixes warnings on line 150 in functions.php

+++ functions.php
@@ -1198,8 +1198,8 @@ if (!function_exists('format_bytes')) {
     function format_bytes($bytes, $precision = 2) {
         $units = array('B', 'KB', 'MB', 'GB', 'TB');
         $bytes = max($bytes, 0);
-        $pow = floor(($bytes ? log($bytes) : 0) / 1024);
-        $pow = min($pow, count($units) - 1);
+    $pow = (int) floor(($bytes ? log($bytes) : 0) / 1024);
+    $pow = min($pow, count($units) - 1);
         $bytes /= (1 << (10 * $pow));
         return round($bytes, $precision) . ' ' . $units[$pow];
     }
